+++ src/modal-confirm3.component.ts
@@ -3,7 +3,7 @@ import { ModalConfirmComponent } from './modal-confirm.component';
 
 @Component({
     selector: 'modal-confirm3',
-    template: `<modal (onOpen)="onOpen()" (onClose)="onCloseModal()" [routed]="false" [isNotification]="isNotification" [settings]="settings">
+    template: `<modal (closemodal)="onCloseModal()" [routed]="false" [isNotification]="isNotification" [settings]="settings">
     <modal-header>
         <h2>{{ title }}</h2>
     </modal-header>
@@ -16,7 +16,7 @@ import { ModalConfirmComponent } from './modal-confirm.component';
         <div [class]="options.confirmFooterToolbarClass">
             <ul [class]="options.confirmFooterButtonsClass">
                 <li [class]="options.confirmFooterButtonItemClass">
-                    <button role="cancel" type="button" [class]="options.confirmCancelButtonClass" (click)="cancel()" #confirmCancel data-dismiss="modal">{{ cancelLabel }}</button>
+                    <button role="cancel" type="button" [class]="options.confirmCancelButtonClass" (click)="cancel()" data-dismiss="modal">{{ cancelLabel }}</button>
                 </li>
                 <li [class]="options.confirmFooterButtonItemClass">
                     <button role="ok" type="button" [class]="options.confirmOkayButtonClass" (click)="ok()">{{ okayLabel }}</button>